htcondorcern: fix Docker network settings
Sets Docker networking strategy to 'host' so that jobs would have better
access to outside sites.

Sets Requirements for the HTCondor version to be 8.9.11, so that jobs
are ensured to end up on machines that will pay attention to the
ClassAd. (This fix should not be necessary in a few months after the
CERN HTCondor update.)

This fixes the network access troubles in workflows such as
`reana-demo-atlas-recast`.